requirements.txt
# Python зависимости для MIRA
# Установка: pip install -r requirements.txt

# ECMWF API Client
ecmwf-api-client>=1.6.0
